Refactor combat_ui.py, restore some codes
